bug fix h12 heroku error
