+++ logs.txt
@@ -1,5 +1,2 @@
-GET		/api/v1/on-covid-19/		200		223224 ms
-GET		/api/v1/on-covid-19/		200		235285 ms
-POST		/api/v1/on-covid-19		200		254393 ms
-POST		/api/v1/on-covid-19		200		816996 ms
-GET		/api/v1/on-covid-19		200		903698 ms
+GET		/api/v1/on-covid-19/		200		20809 ms
+GET		/setup		200		28380 ms
